Spelling Error Fix & Fixing Cover Images
Fixed a spelling error early in the code (small, but I couldn't commit code and not suggest it)

I couldn't get the covers to work at all - I ended up running the debugger and found where it removed it. I added it in so that the image now works properly.