Update ASIN, add Duration and Date Published
The previous line for getting the ASIN was not always present, and would cause a failure of the entire tagging process.  

Added Duration so that I can compare the expected length against the actual file length to be sure the book is a match and there are no missing chapters, etc. 

Added Date Published.  I will be substituting this in for the YEAR tag in an mp3tag action.   I would change the script to put it in directly, but do not want to presume this is wanted by everyone.